remove env storting for hash calculation
